channel_template/main_automation.py
"""
Generic YouTube Automation - Main Entry Point
Use this script to run the full pipeline for any niche.
"""

import os
import sys
import yaml
import argparse
from datetime import datetime
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class GenericChannelAutomation:
    def __init__(self, config_path="config/config.yaml"):
        # Load configuration
        if not os.path.exists(config_path):
            raise FileNotFoundError(f"Config file not found: {config_path}. Please rename config.yaml.example first.")
            
        with open(config_path, 'r', encoding='utf-8') as f:
            self.config = yaml.safe_load(f)
            
        print(f"🚀 Initializing '{self.config['content']['niche']}' Automation...")
        
        # Initialize modules (Placeholders for now)
        # form modules.script_engine import ScriptEngine
        # from modules.media_engine import MediaEngine
        # self.script_engine = ScriptEngine(self.config)
        # self.media_engine = MediaEngine(self.config)
        
    def run_daily_job(self):
        print("📅 Running daily content generation...")
        # 1. Generate Script
        # script = self.script_engine.generate_script()
        
        # 2. Get Media
        # media = self.media_engine.get_footage(script['visual_keywords'])
        
        # 3. Create Video (Placeholder)
        print("✅ Daily job complete (Template Mode)")

if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="Run YouTube Automation")
    parser.add_argument("--test", action="store_true", help="Run a test generation")
    args = parser.parse_args()
    
    automation = GenericChannelAutomation()
    automation.run_daily_job()
