channel_template/requirements.txt
openai>=1.12.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pyyaml>=6.0
google-api-python-client>=2.115.0
moviepy<2.0.0
requests>=2.31.0
elevenlabs>=0.2.26
